tools/PeakRDL-repl/pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "peakrdl-repl"
license-files = ["LICENSE"]
dynamic = ["version"]
requires-python = ">=3.11"
dependencies = [
    "systemrdl-compiler >= 1.21.0, < 2",
]

authors = [
    {name="Antmicro"},
]

description = "Generate Renode REPL platform files from a SystemRDL model"
readme = "README.md"

keywords = [
    "SystemRDL", "PeakRDL", "CSR", "compiler", "tool", "platform", "generator",
    "repl", "renode", "header", "software",
]

[project.urls]
Source = "https://github.com/renode/renode/tree/master/tools/PeakRDL-repl"
Tracker = "https://github.com/renode/renode/issues"

[tool.setuptools.dynamic]
version = {attr = "peakrdl_repl.__about__.__version__"}

[project.entry-points."peakrdl.exporters"]
renode-repl = "peakrdl_repl.__peakrdl__:ReplExporterDescriptor"
